Added "Mapping.java" as well as three test icons that I added as resources. Unfortunately Mapping doesn't like running using those images as buttons. I'll fix it later, but atleast you can see what it looks like in the previewer. These icons are really place-holders for the proper ones I'll make later on for it.
I've got most of the text set to one shade lighter than black so it gives a bit of a softer feel to it. This doesn't have to be replicated throughout the rest of the program, it's just something I like to do. The panel at the very bottom is a status bar of sorts; it'll give updates on what the program is doing and other miscellaneous information.