Added "Mapping.java". The icons are http://i45.tinypic.com/cjpkg.png and http://i49.tinypic.com/yigr6.jpg  I tried adding the images as resources, but it gave me errors every time I tried to run it. These are really just place holders until I make a set of proper icons. I may nitpick at the sizes and alignment of the objects in this frame, but this is pretty much what it will be.
I've got most of the text set to one shade lighter than black so it gives a bit of a softer feel to it. This doesn't have to be replicated throughout the rest of the program, it's just something I like to do. The panel at the very bottom is a status bar of sorts; it'll give updates on what the program is doing and other miscellaneous information.